Add swipe to learn/unlearn words
This commit adds the ability to set specific words as learned or unlearn
individual words. Currently you can reset categories, but not individual
words. This is useful if you're looking back at a word and you realize
you've forgotten it, so you want to review it again. To trigger this
action, you swipe left on the word. The ability to swipe right on the
word to mark it as learned is in case you accidentally left swipe the
wrong word.